backend/requirements.txt
flask
flask-cors
requests
python-dotenv
werkzeug
deep-translator
gunicorn
